% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idig_search.R
\name{build_field_lists}
\alias{build_field_lists}
\title{Build fields and fields_exclude for queries.}
\usage{
build_field_lists(fields, type)
}
\arguments{
\item{fields}{character vector of fields user wants returned}

\item{type}{type of records to get fields for}
}
\value{
list list with fields key for df fields and query key for parameters
to be merged with the query sent
}
\description{
Given the desired fields to be returned, intelligently add an exclusion for
the data array if warranted and handle the "all" keyword. And do so without
setting both fields and fields_exclude due to fact that the API will return
wrong results if are passed. This is still posssible if the user
deliberately sets both. Not exported.
}

