% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idig_build_attrib.R
\name{idig_build_attrib}
\alias{idig_build_attrib}
\title{Attribution dataframe of iDigBio records query}
\usage{
idig_build_attrib(dat)
}
\arguments{
\item{dat}{dataframe generated by idig_search method}
}
\value{
a data frame
}
\description{
Function to build attribution dataframe from a query to the iDigBio API
}
\details{
This function differs from the attribution metadata that is attached to the
dataframe returned by the idig_search_* methods. It summarizes the record 
sets used by records in the dataframe, not the record sets that have records
that match the query sent to iDigBio. This is useful if only part of the 
records for a query are downloaded, for example with the limit and offset
parameters.

Exported.
}
\author{
Kevin Love
}

