% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{idig_check}
\alias{idig_check}
\title{check HTTP code}
\usage{
idig_check(req)
}
\arguments{
\item{req}{the returned request}
}
\value{
nothing. Stops if HTTP code is < 400
}
\description{
checks for HTTP codes
}
\details{
Part 1 of the error checking process. not exported.
}
\author{
Francois Michonneau
}

