% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{idig_check_error}
\alias{idig_check_error}
\title{Check is the request returned an error.}
\usage{
idig_check_error(req)
}
\arguments{
\item{req}{the returned request}
}
\value{
nothing. Stops if request contains an error.
}
\description{
checks for error messages that can be returned by the API
}
\details{
Part 2 of the error checking process. not exported.
}
\author{
Francois Michonneau
}

