% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idig_count_media.R
\name{idig_count_media}
\alias{idig_count_media}
\title{Count media endpoint}
\usage{
idig_count_media(rq = FALSE, mq = FALSE, ...)
}
\arguments{
\item{rq}{iDigBio record query in nested list format}

\item{mq}{iDigBio media query in nested list format}

\item{...}{additional parameters}
}
\value{
count of media records matching the query(s)
}
\description{
Count media records matching a query.
}
\details{
Quickly return a count of the media records matching the query(s) provided.
}
\author{
Matthew Collins
}

