% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idig_count_records.R
\name{idig_count_records}
\alias{idig_count_records}
\title{count record endpoint}
\usage{
idig_count_records(rq = FALSE, ...)
}
\arguments{
\item{rq}{iDigBio record query in nested list format}

\item{...}{additional parameters}
}
\value{
count of specimen records matching the query(s)
}
\description{
Count specimen records matching a query.
}
\details{
Quickly return a count of the specimen records matching the query(s)
provided.
}
\author{
Matthew Collins
}

