% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{idig_GET}
\alias{idig_GET}
\title{internal GET request}
\usage{
idig_GET(path, ...)
}
\arguments{
\item{path}{endpoint}

\item{...}{additional arguments to be passed to httr::GET}
}
\value{
the request (as a list)
}
\description{
internal function for GET requests
}
\details{
Generates a GET request and performs the checks on what is
returned.
}
\author{
Francois Michonneau
}

