% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idig_meta_fields.R
\name{idig_meta_fields}
\alias{idig_meta_fields}
\title{meta fields endpoint}
\usage{
idig_meta_fields(type = "records", subset = FALSE, ...)
}
\arguments{
\item{type}{string type of fields to return, defaults to "records"}

\item{subset}{set of fields to return, "indexed", "raw", or unset for all}

\item{...}{additional parameters}
}
\value{
list of fields of the requested type
}
\description{
List of fields in iDigBio.
}
\details{
Return a list of media or specimen fields that are contained in iDigBio.
}
\author{
Matthew Collins
}

