% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{idig_parse}
\alias{idig_parse}
\title{parse successfully returned request}
\usage{
idig_parse(req)
}
\arguments{
\item{req}{the returned request}
}
\value{
a list
}
\description{
Parses output of successful query to return a list.
}
\details{
Not exported.
}
\author{
Francois Michonneau
}
