% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{idig_POST}
\alias{idig_POST}
\title{internal POST request}
\usage{
idig_POST(path, body, encode = "json", ...)
}
\arguments{
\item{path}{endpoint}

\item{body}{a named list inside a list named "rq"}

\item{encode}{the API wants "json"}

\item{...}{additional arguments to be passed to httr::POST}
}
\value{
the request (as a list)
}
\description{
internal function for POST requests
}
\details{
Generates a POST request and performs the checks on what is
returned
}
\author{
Francois Michonneau
}

