% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{idig_search}
\alias{idig_search}
\title{Basic searching of the iDigBio API}
\usage{
idig_search(query, limit = 100, offset = 0,
  fields = c("dwc:catalogNumber", "dwc:genus", "dwc:specificEpithet",
  "dwc:decimalLatitude", "dwc:decimalLongitude"), ...)
}
\arguments{
\item{query}{a list containing the information to be searched in iDigBio}

\item{limit}{maximum number of records to return}

\item{offset}{where to search the search}

\item{fields}{list of fields that will be contained in the data.frame}

\item{...}{additional parameters (currently not in use)}
}
\value{
a data frame
}
\description{
Function to query the iDigBio API
}
\details{
Currently query needs to be specified as a list
}
\examples{
\dontrun{
idig_search(query=list(genus="acer"), limit=10)
}
}
\author{
Francois Michonneau
}

