% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{idig_url}
\alias{idig_url}
\title{base URL}
\usage{
idig_url(dev = TRUE)
}
\arguments{
\item{dev}{Should be the beta version of the API be used?}
}
\value{
string for the URL
}
\description{
base URL for the API
}
\details{
Not exported
}
\author{
Francois Michonneau
}

