% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{idig_validate}
\alias{idig_validate}
\title{Stub function for validating parameters.}
\usage{
idig_validate(inputs)
}
\description{
Takes list of inputs named by validation rule eg "number":[2, 3] and returns
a vector of strings with any validation errors. If the vector is 0 length,
everything is valid.
}

