% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idig_view_media.R
\name{idig_view_media}
\alias{idig_view_media}
\title{view media endpoint}
\usage{
idig_view_media(uuid, ...)
}
\arguments{
\item{uuid}{uuid of media record}

\item{...}{additional parameters}
}
\value{
nested list of data
}
\description{
View individual media records.
}
\details{
View all information about a specific media record.
}
\author{
Matthew Collins
}

