% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idig_view_records.R
\name{idig_view_records}
\alias{idig_view_records}
\title{view specimen endpoint}
\usage{
idig_view_records(uuid, ...)
}
\arguments{
\item{uuid}{uuid of specimen record}

\item{...}{additional parameters}
}
\value{
nested list of data
}
\description{
View individual specimen records.
}
\details{
View all information about a specific specimen record.
}
\author{
Matthew Collins
}
