% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridigbio.R
\name{ridigbio}
\alias{ridigbio}
\title{Retrieve data from the iDigBio specimen data repository.}
\description{
Retrieve data from the iDigBio specimen data repository.
}
\section{About}{


ridigbio provides an interface to the iDigBio data API described here: 
\url{https://www.idigbio.org/wiki/index.php/IDigBio_API}. With this package
you can retrieve specimen and media records from the iDigBio data 
repository. The iDigBio portal \url{https://portal.idigbio.org/} uses the 
same API so you should be able to retrieve the same information as shown in
the portal.

iDigBio contains nearly 30 million data records on museum specimens held at
United States institutions. It also holds nearly 5 million images of these
specimens.
}

\section{API Updates}{

API Update as of 0.3.8:

Default fields have been changed to accomodate data users regarding dates found in the field "datecollected".
We will no longer be providing this field to users by default and will instead be returning the following fields:

data.dwc:eventDate
data.dwc:year
data.dwc:month
data.dwc:day

Please be advised that these fields are in a text only format and will need to be sorted another way as addressed in:
\url{https://github.com/iDigBio/ridigbio/issues/44}

}

\section{Getting Started}{


The main function is \code{\link{idig_search_records}} and reviewing its
documenation first with \code{?idig_search_records} is recommended.
}

\section{Limitations}{


This package does not yet provide an interface to the mapping or the 
download APIs.
}

\section{Citing}{


To cite the ridigbio package in your work, please use the following format:

Michonneau F, Collins M, Chamberlain SA (2016). ridigbio: An interface to iDigBio's search API that allows downloading specimen records. R package version 0.3.8. https://github.com/iDigBio/ridigbio
}

\author{
Francois Michonneau \email{francois.michonneau@gmail.com}

Matthew Collins \email{mcollins@acis.ufl.edu}
}
