% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{idig_check}
\alias{idig_check}
\title{check HTTP code}
\usage{
idig_check(req)
}
\arguments{
\item{req}{the returned request}
}
\value{
nothing. Stops if HTTP code is < 400
}
\description{
checks for HTTP codes
}
\details{
Part 1 of the error checking process. Not exported.
}
\author{
Francois Michonneau
}

