% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{idig_check}
\alias{idig_check}
\title{check HTTP code}
\usage{
idig_check(req)
}
\arguments{
\item{req}{the returned request}
}
\value{
nothing. Stops if HTTP code is >= 400
}
\description{
Checks for HTTP error codes and JSON errors.
}
\details{
Part 1 of the error checking process. This part handles HTTP error codes and
then calls part 2 which handles JSON errors in the responses. Not exported.
}
\author{
Francois Michonneau
}

