% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{idig_GET}
\alias{idig_GET}
\title{internal GET request}
\usage{
idig_GET(path, ...)
}
\arguments{
\item{path}{endpoint}

\item{...}{additional arguments to be passed to httr::GET}
}
\value{
the request (as a list)
}
\description{
Internal function for GET requests.
}
\details{
Generates a GET request and performs the checks on what is returned. Not
exported.
}
\author{
Francois Michonneau
}
