% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{idig_parse}
\alias{idig_parse}
\title{parse successfully returned request}
\usage{
idig_parse(req)
}
\arguments{
\item{req}{the returned request}
}
\value{
a list
}
\description{
parses output of successful query to return a list
}
\details{
not exported
}
\author{
Francois Michonneau
}

