% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{idig_POST}
\alias{idig_POST}
\title{internal POST request}
\usage{
idig_POST(path, body, ...)
}
\arguments{
\item{path}{endpoint}

\item{body}{a list of parameters for the endpoint}

\item{...}{additional arguments to be passed to httr::POST}
}
\value{
the request (as a list)
}
\description{
Internal function for POST requests.
}
\details{
Generates a POST request and performs the checks on what is returned. Not
exported.
}
\author{
Francois Michonneau
}
