% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idig_search_media.R
\name{idig_search_media}
\alias{idig_search_media}
\title{Basic searching of iDigBio media}
\usage{
idig_search_media(mq = FALSE, rq = FALSE, fields = FALSE,
  max_items = 1e+05, limit = 0, offset = 0, sort = FALSE)
}
\arguments{
\item{fields}{list of fields that will be contained in the data.frame}

\item{max_items}{maximum number of results allowed to be retrieved}

\item{limit}{maximum number of results returned}

\item{offset}{number of results to skip before returning results}

\item{query}{a list containing the information to be searched in iDigBio}

\item{...}{additional parameters (currently not in use)}
}
\value{
a data frame
}
\description{
Function to query the iDigBio API for specimen media
}
\details{
Currently the query needs to be specified as a list. All matching results are
returned up to the max_items cap (default 100,000). If more results are
wanted, the max_items can be passed as an option. Limit and offset are
availible if manual paging of results is needed though the max_items cap
still applies as the item count comes from the results header not the
count of actual media in the limit/offset window.

Return is a data.frame containing the requested fields (or the dfault fields).
Only fields from the Elasticsearch index are currently availible, no raw
fields. As such, the columns in the data frame are types however no factors
are built. Attribution and other metadata is attached to the dataframe in the
data.frame's attributes. (I.e. attributes(df))
}
\examples{
\dontrun{
idig_search(rq=list(genus="acer"), limit=10)
}
}
\author{
Francois Michonneau
}

