% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idig_search_records.R
\name{idig_search_records}
\alias{idig_search_records}
\title{Basic searching of iDigBio records}
\usage{
idig_search_records(rq, fields = FALSE, max_items = 1e+05, limit = 0,
  offset = 0, sort = FALSE, ...)
}
\arguments{
\item{rq}{iDigBio record query in nested list format}

\item{fields}{vector of fields that will be contained in the data.frame,
limited set returned by default, use "all" to get all indexed fields}

\item{max_items}{maximum number of results allowed to be retrieved (fail
-safe)}

\item{limit}{maximum number of results returned}

\item{offset}{number of results to skip before returning results}

\item{sort}{vector of fields to use for sorting, if paging always include
UUID to get reliable record order}

\item{...}{additional parameters}
}
\value{
a data frame
}
\description{
Function to query the iDigBio API for specimen records
}
\details{
Currently the query needs to be specified as a list. All matching results
are returned up to the max_items cap (default 100,000). If more results are
wanted, the max_items can be passed as an option. Limit and offset are
availible if manual paging of results is needed though the max_items cap
still applies as the item count comes from the results header not the
count of actual records in the limit/offset window.

Return is a data.frame containing the requested fields (or the default
fields). The columns in the data frame are types however no factors are
built. Attribution and other metadata is attached to the dataframe in the
data.frame's attributes. (I.e. attributes(df)) Not exported.
}
\examples{
\dontrun{
idig_search(rq=list(genus="acer"), limit=10)
}
}
\author{
Matthew Collins
}

