% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{idig_url}
\alias{idig_url}
\title{base URL}
\usage{
idig_url(dev = FALSE)
}
\arguments{
\item{dev}{Should be the beta version of the API be used?}
}
\value{
string for the URL
}
\description{
Return base URL for the API calls.
}
\details{
Defaults to use beta URL. Not exported.
}
\author{
Francois Michonneau
}

