% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{idig_validate}
\alias{idig_validate}
\title{validate fields}
\usage{
idig_validate(inputs)
}
\arguments{
\item{inputs}{list of inputs to validate}
}
\value{
boolean
}
\description{
Stub function for validating parameters.
}
\details{
Takes list of inputs named by validation rule eg "number":[2, 3] and returns
a vector of strings with any validation errors. If the vector is 0 length, 
everything is valid. Not exported.
}
\author{
Matthew Collins
}

