% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{idig_version}
\alias{idig_version}
\title{API version}
\usage{
idig_version(version = "v2")
}
\arguments{
\item{version}{optional argument giving the version of the API to use}
}
\value{
string for the version to use
}
\description{
version number to use for the API
}
\details{
current default is "v2". Function not exported.
}
\author{
Francois Michonneau
}

