% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{idig_version}
\alias{idig_version}
\title{API version}
\usage{
idig_version(version = "v2")
}
\arguments{
\item{version}{optional argument giving the version of the API to use}
}
\value{
string for the version to use
}
\description{
Return the version number to use for the API calls.
}
\details{
The current default is "v2". Not exported.
}
\author{
Francois Michonneau
}
