% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridigbio-package.R
\docType{package}
\name{ridigbio-package}
\alias{ridigbio-package}
\alias{_PACKAGE}
\title{ridigbio: Interface to the iDigBio Data API}
\description{
An interface to iDigBio's search API that allows downloading specimen records. Searches are returned as a data.frame. Other functions such as the metadata end points return lists of information. iDigBio is a US project focused on digitizing and serving museum specimen collections on the web. See \url{https://www.idigbio.org} for information on iDigBio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iDigBio/ridigbio}
  \item Report bugs at \url{https://github.com/iDigBio/ridigbio/issues}
}

}
\author{
\strong{Maintainer}: Jesse Bennett \email{j3553@ufl.edu} (Maintainer)

Authors:
\itemize{
  \item Francois Michonneau \email{francois.michonneau@gmail.com} (Original Author) [copyright holder]
  \item Matthew Collins \email{mcollins@acis.ufl.edu} (Original Author)
}

Other contributors:
\itemize{
  \item Scott Chamberlain \email{myrmecocystus@gmail.com} [contributor]
  \item Kevin Love \email{klove@flmnh.ufl.edu} [contributor]
  \item Hem Nalini Morzaria-Luna \email{hemnalin.morzarialuna@noaa.gov} [contributor]
  \item Shelly Gaynor \email{michellegaynor@ufl.edu} [contributor]
}

}
\keyword{internal}
